requirements.txt
Django
djangorestframework
psycopg2-binary
celery
redis
openpyxl
python-decouple
django-cors-headers